+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/FileBased/Parser/ParserResourceTests.cs
@@ -66,12 +66,12 @@ public class ParserResourceTests
 
 #if NET
         string[] expectedDetecors = [
-            "azureappservice", "container", "host", "operatingsystem", "process", "processruntime"
+            "azureappservice", "azurecontainerapps", "container", "host", "operatingsystem", "process", "processruntime"
                 ];
 #endif
 #if NETFRAMEWORK
         string[] expectedDetecors = [
-            "azureappservice", "host", "operatingsystem", "process", "processruntime"
+            "azureappservice", "azurecontainerapps", "host", "operatingsystem", "process", "processruntime"
                 ];
 #endif