+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   OpAMP custom messages.
 - Support for [StackExchange.Redis](https://www.nuget.org/packages/StackExchange.Redis)
   traces instrumentation for versions `3.1.x`+ for .NET only.
+- `AZURECONTAINERAPPS` resource detector, detecting Azure Container Apps
+  resource information.
 
 ### Changed